Add documentation for prepared parameters + make it eaiser to use
